testing

+++ src/main/java/com/ls/backoffice/service/PostService.java
@@ -1,6 +1,7 @@
 package com.ls.backoffice.service;
 
 import com.ls.backoffice.domain.model.Post;
+import com.ls.backoffice.exception.NotExistByIdExeption;
 import com.ls.backoffice.repository.IPostRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,12 +25,19 @@ public class PostService implements IPostService {
     @Override
     public Post getPost(String postId) {
         log.info("getting post by Id");
-        return postRepository.findById(postId).orElseGet(()->null);
+        return postRepository.findById(postId)
+                .orElseThrow(()->new NotExistByIdExeption("El post solicitado no existe"));
+
     }
 
     @Override
     public Post updatePost(Post post) {
-        return postRepository.save(post);
+        if (postRepository.existsById(post.getPostId())) {
+            return postRepository.save(post);
+        }
+        else{
+            throw new NotExistByIdExeption("El post solicitado no existe");
+        }
     }
 
     @Override
@@ -39,6 +47,11 @@ public class PostService implements IPostService {
 
     @Override
     public void deletePost(String postId) {
-        postRepository.deleteById(postId);
+        if (postRepository.existsById(postId)) {
+            postRepository.deleteById(postId);
+        }
+        else{
+            throw new NotExistByIdExeption("El post solicitado no existe");
+        }
     }
 }
